pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf2md"
dynamic = ["version"]
description = "CLI tool that converts PDF documents to well-structured Markdown"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pymupdf4llm>=0.0.17",
    "click>=8.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3",
    "pdf2image>=1.16",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "pyinstaller>=6.0",
]

[project.scripts]
pdf2md = "pdf2md.cli:main"

[tool.hatch.version]
path = "pdf2md/__init__.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
